+++ mobile/android/android-components/components/browser/session/src/test/java/mozilla/components/browser/session/SessionTest.kt
@@ -13,6 +13,7 @@ import mozilla.components.browser.session.Session.Source
 import mozilla.components.browser.session.manifest.WebAppManifest
 import mozilla.components.browser.session.tab.CustomTabConfig
 import mozilla.components.concept.engine.HitResult
+import mozilla.components.concept.engine.media.Media
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.window.WindowRequest
@@ -36,6 +37,7 @@ import org.mockito.Mockito.spy
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
 import org.mockito.Mockito.verifyNoMoreInteractions
+import java.lang.IllegalArgumentException
 
 class SessionTest {
     @Test
@@ -605,6 +607,8 @@ class SessionTest {
         defaultObserver.onOpenWindowRequested(session, windowRequest)
         defaultObserver.onCloseWindowRequested(session, windowRequest)
         defaultObserver.onWebAppManifestChanged(session, mock())
+        defaultObserver.onMediaAdded(session, emptyList(), mock())
+        defaultObserver.onMediaRemoved(session, emptyList(), mock())
     }
 
     @Test
@@ -785,5 +789,98 @@ class SessionTest {
     fun `toString returns string containing id and url`() {
         val session = Session(id = "my-session-id", initialUrl = "https://www.mozilla.org")
         assertEquals("Session(my-session-id, https://www.mozilla.org)", session.toString())
+    }
+
+    @Test
+    fun `observer is notified when media is added`() {
+        var observedList: List<Media>? = null
+        var observedMedia: Media? = null
+
+        val observer = object : Session.Observer {
+            override fun onMediaAdded(session: Session, media: List<Media>, added: Media) {
+                observedList = media
+                observedMedia = added
+            }
+        }
+
+        val session = Session("https://www.mozilla.org")
+        session.register(observer)
+
+        val addedMedia1: Media = mock()
+        session.media = listOf(addedMedia1)
+
+        assertEquals(addedMedia1, observedMedia)
+        assertEquals(listOf(addedMedia1), observedList)
+
+        val addedMedia2: Media = mock()
+        session.media = listOf(addedMedia1, addedMedia2)
+
+        assertEquals(addedMedia2, observedMedia)
+        assertEquals(listOf(addedMedia1, addedMedia2), observedList)
+
+        val addedMedia3: Media = mock()
+        session.media = listOf(addedMedia1, addedMedia3, addedMedia2)
+
+        assertEquals(addedMedia3, observedMedia)
+        assertEquals(listOf(addedMedia1, addedMedia3, addedMedia2), observedList)
+    }
+
+    @Test
+    fun `observer is notified when media is removed`() {
+        var observedList: List<Media>? = null
+        var observedMedia: Media? = null
+
+        val observer = object : Session.Observer {
+            override fun onMediaRemoved(session: Session, media: List<Media>, removed: Media) {
+                observedList = media
+                observedMedia = removed
+            }
+        }
+
+        val media1: Media = mock()
+        val media2: Media = mock()
+        val media3: Media = mock()
+
+        val session = Session("https://www.mozilla.org")
+        session.media = listOf(media1)
+        session.media = listOf(media1, media2)
+        session.media = listOf(media1, media2, media3)
+        session.register(observer)
+
+        session.media = listOf(media1, media2)
+
+        assertEquals(media3, observedMedia)
+        assertEquals(listOf(media1, media2), observedList)
+
+        session.media = listOf(media2)
+
+        assertEquals(media1, observedMedia)
+        assertEquals(listOf(media2), observedList)
+
+        session.media = listOf()
+
+        assertEquals(media2, observedMedia)
+        assertEquals(emptyList<Media>(), observedList)
+    }
+
+    @Test(expected = IllegalArgumentException::class)
+    fun `Session throws if more than one Media object is added at a time`() {
+        val session = Session("https://www.mozilla.org")
+        session.media = listOf(mock(), mock())
+    }
+
+    @Test(expected = IllegalArgumentException::class)
+    fun `Session throws if more than one Media object is removed at a time`() {
+        val session = Session("https://www.mozilla.org")
+
+        val media1: Media = mock()
+        val media2: Media = mock()
+        val media3: Media = mock()
+
+        session.media = listOf(media1)
+        session.media = listOf(media1, media2)
+        session.media = listOf(media1, media2, media3)
+
+        session.media = listOf(media1)
     }
 }